+++ sam/metrics/tests/test_incident_recall.py
@@ -75,7 +75,9 @@ class TestIncidentPrecisionRecallCurve(unittest.TestCase):
         y_incidents = [0, 0, 0, 1]
         y_pred = [0.1, 0.2, 0.3, 0.4]
         p, r, t = precision_incident_recall_curve(y_incidents, y_pred, range_pred=(0, 1))
-        assert_array_almost_equal(p, np.array([0.5, 0.666667, 1, 1, 1]))
+        # Note: Behaviour of sklearn precision recall curve changed in version 1.1.1
+        # Now the number of thresholds = number of unique predictions
+        assert_array_almost_equal(p, np.array([0.5, 0.6666666666666666, 1, 1, 1]))
         assert_array_equal(r, np.array([1, 1, 1, 1, 0]))
         assert_array_equal(t, np.array([0.1, 0.2, 0.3, 0.4]))