+++ sam/validation/__init__.py
@@ -1,5 +1,15 @@
-from .find_extremes import RemoveExtremeValues
-from .find_flatlines import RemoveFlatlines
+from .base_validator import BaseValidator
+from .outside_range_validator import OutsideRangeValidator
+from .mad_validator import RemoveExtremeValues, MADValidator
+from .flatline_validator import RemoveFlatlines, FlatlineValidator
 from .setup_validation_pipeline import create_validation_pipe
 
-__all__ = ["RemoveExtremeValues", "RemoveFlatlines", "create_validation_pipe"]
+__all__ = [
+    "BaseValidator",
+    "OutsideRangeValidator",
+    "RemoveExtremeValues",
+    "MADValidator",
+    "RemoveFlatlines",
+    "FlatlineValidator",
+    "create_validation_pipe",
+]